refactor: separate update queue from DOM and fix architectural layers
