+++ packages/web-components/fast-element/src/templating/children.spec.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import { children, ChildrenDirective } from "./children";
 import { ExecutionContext, observable } from "../observation/observable";
 import { elements } from "./node-observation";
-import { DOM } from "../dom";
+import { Updates } from "../observation/update-queue";
 
 describe("The children", () => {
     context("template function", () => {
@@ -90,7 +90,7 @@ describe("The children", () => {
 
             const updatedChildren = children.concat(createAndAppendChildren(host));
 
-            await DOM.nextUpdate();
+            await Updates.next();
 
             expect(model.nodes).members(updatedChildren);
         });
@@ -110,7 +110,7 @@ describe("The children", () => {
 
             const updatedChildren = children.concat(createAndAppendChildren(host));
 
-            await DOM.nextUpdate();
+            await Updates.next();
 
             expect(model.nodes).members(updatedChildren.filter(elements("foo-bar")));
         });
@@ -151,7 +151,7 @@ describe("The children", () => {
                 }
             }
 
-            await DOM.nextUpdate();
+            await Updates.next();
 
             expect(model.nodes).members(subtreeChildren);
         });
@@ -174,7 +174,7 @@ describe("The children", () => {
 
             host.appendChild(document.createElement("div"));
 
-            await DOM.nextUpdate();
+            await Updates.next();
 
             expect(model.nodes).members([]);
         });
